testplat/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django==2.2.5
django-celery==3.2.2
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
kombu==3.0.37
MarkupSafe==1.1.1
mysqlclient==1.4.4
openapi-codec==1.3.2
Pillow==6.1.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.8
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
